All painters done

+++ src/main/java/raf/draft/dsw/controller/states/AbstractState.java
@@ -4,9 +4,9 @@ import raf.draft.dsw.gui.swing.mainpanel.room.tab.RoomTab;
 import raf.draft.dsw.model.structures.room.interfaces.VisualElement;
 
 public abstract class AbstractState {
-    void mouseClick(int x, int y, VisualElement element, RoomTab roomTab){}
-    void mousePressed(int x, int y, VisualElement element, RoomTab roomTab){}
-    void mouseReleased(int x, int y, VisualElement element, RoomTab roomTab){}
-    void mouseDragged(int dx, int dy, VisualElement element, RoomTab roomTab){}
-    void mouseWheelScrolled(int x, int y, double wheelRotation, RoomTab roomTab){}
+    void mouseClick(double x, double y, VisualElement element, RoomTab roomTab){}
+    void mousePressed(double x, double y, VisualElement element, RoomTab roomTab){}
+    void mouseReleased(double x, double y, VisualElement element, RoomTab roomTab){}
+    void mouseDragged(double dx, double dy, VisualElement element, RoomTab roomTab){}
+    void mouseWheelScrolled(double x, double y, double wheelRotation, RoomTab roomTab){}
 }
